Implement code review changes
